+++ MonsterChase/MonsterChase/ChasingMonster.cpp
@@ -41,16 +41,16 @@ void ChasingMonster::move_chasing(const Vector2D &target_pos) {
 	float game_object_x = game_object->get_position().x();
 	float game_object_y = game_object->get_position().y();
 	if (game_object_x > target_x) {
-		game_object->set_position(Vector2D(game_object_x - 1.0, game_object_y));
+		game_object->set_position(Vector2D(game_object_x - static_cast<float>(1.0), game_object_y));
 	}
 	else if (game_object_x < target_x) {
-		game_object->set_position(Vector2D(game_object_x + 1.0, game_object_y));
+		game_object->set_position(Vector2D(game_object_x + static_cast<float>(1.0), game_object_y));
 	}
 	if (game_object_y > target_y) {
-		game_object->set_position(Vector2D(game_object_x, game_object_y - 1.0));
+		game_object->set_position(Vector2D(game_object_x, game_object_y - static_cast<float>(1.0)));
 	}
 	else if (game_object_y < target_y) {
-		game_object->set_position(Vector2D(game_object_x, game_object_y + 1.0));
+		game_object->set_position(Vector2D(game_object_x, game_object_y + static_cast<float>(1.0)));
 	}
 }
